+++ addons/next-auth/api-handler.ts
@@ -1,13 +1,8 @@
 import NextAuth from "next-auth";
 import GithubProvider from "next-auth/providers/github";
 
-// Prisma adapter for NextAuth, optional and can be removed
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import type { PrismaClient } from "@prisma/client";
-
 export default NextAuth({
   // Configure one or more authentication providers
-  adapter: PrismaAdapter(prisma as PrismaClient),
   providers: [
     GithubProvider({
       clientId: process.env.GITHUB_ID,
