+++ src/installers/tailwind.ts
@@ -1,0 +1,7 @@
+// Steps: https://tailwindcss.com/docs/guides/nextjs
+
+// 1. Install dependencies
+// pnpx i -D tailwindcss postcss autoprefixer
+
+// 2. Copy getting-started files
+// addons/tailwind/*
