+++ src/installers/trpc.ts
@@ -1,0 +1,12 @@
+// Install dependencies
+
+/**
+ * @trpc/server
+ * @trpc/next
+ * @trpc/client
+ * @trpc/react
+ * zod
+ * superjson
+ */
+
+// Copy getting-started files
