--- template-prisma-auth/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "create-t3-app",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@next-auth/prisma-adapter": "^1.0.3",
-    "@prisma/client": "3.14.0",
-    "@trpc/client": "^9.23.4",
-    "@trpc/next": "^9.23.4",
-    "@trpc/react": "^9.23.4",
-    "@trpc/server": "^9.23.4",
-    "next": "12.1.6",
-    "next-auth": "^4.5.0",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-query": "^3.39.0",
-    "superjson": "^1.9.1",
-    "zod": "^3.16.0"
-  },
-  "devDependencies": {
-    "@types/node": "17.0.35",
-    "@types/react": "18.0.9",
-    "@types/react-dom": "18.0.4",
-    "autoprefixer": "^10.4.7",
-    "eslint": "8.15.0",
-    "eslint-config-next": "12.1.6",
-    "postcss": "^8.4.14",
-    "prisma": "^3.14.0",
-    "tailwindcss": "^3.0.24",
-    "typescript": "4.6.4"
-  }
-}
