--- template-prisma-auth/src/server/create-router.ts
@@ -1,4 +1,0 @@
-import { router } from "@trpc/server";
-import { Context } from "./context";
-
-export const createRouter = () => router<Context>();
