--- template-prisma/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": ["next/babel", "next/core-web-vitals"]
-}
