--- template-prisma/src/utils/trpc.ts
@@ -1,4 +1,0 @@
-import type { AppRouter } from "../pages/api/trpc/[trpc]";
-import { createReactQueryHooks } from "@trpc/react";
-
-export const trpc = createReactQueryHooks<AppRouter>();
