--- template-prisma/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  content: ["./src/**/*.{js,ts,jsx,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
